drm: correct misra-c violation 10.4
misra_violation:
[CID 1703445] [CID 1703492] [CID 1703578] [CID 1703592]
[CID 1703689] [CID 2202098] [CID 2309393] [CID 2309394]
[CID 2309395] [CID 2309396] [CID 2309397] [CID 2309398]
[CID 2309399] [CID 2309400] [CID 2309401] [CID 2309402]
[CID 2309403] [CID 2309404] [CID 2309405] [CID 2309406]
[Required] MISRA C-2012 Rule 10.4 violation

[Detail]
update some operand type

Test: build pass and test ok

CR-Id: AUTO00005421

+++ drivers/gpu/drm/mediatek/mtk_disp_ovl.c
@@ -165,11 +165,13 @@ static unsigned int ovl_fmt_convert(struct mtk_disp_ovl *ovl, unsigned int fmt)
 	case DRM_FORMAT_RGB565:
 		return OVL_CON_CLRFMT_RGB565(ovl);
 	case DRM_FORMAT_BGR565:
-		return OVL_CON_CLRFMT_RGB565(ovl) | OVL_CON_BYTE_SWAP;
+		return (unsigned int)OVL_CON_CLRFMT_RGB565(ovl) |
+				     OVL_CON_BYTE_SWAP;
 	case DRM_FORMAT_RGB888:
 		return OVL_CON_CLRFMT_RGB888(ovl);
 	case DRM_FORMAT_BGR888:
-		return OVL_CON_CLRFMT_RGB888(ovl) | OVL_CON_BYTE_SWAP;
+		return (unsigned int)OVL_CON_CLRFMT_RGB888(ovl) |
+				     OVL_CON_BYTE_SWAP;
 	case DRM_FORMAT_RGBX8888:
 	case DRM_FORMAT_RGBA8888:
 		return OVL_CON_CLRFMT_ARGB8888;
